+++ kfp/kfp_ray_components/Dockerfile
@@ -19,7 +19,5 @@ COPY --chown=ray:users kfp_support_lib/ kfp_support_lib/
 RUN cd kfp_support_lib && pip install --no-cache-dir -e .
 # remove credentials-containing file
 RUN rm requirements.txt
-
-
-
+# components
 COPY ./src /pipelines/component/src
