+++ kfp/kfp_ray_components/src/execute_ray_job.py
@@ -10,61 +10,8 @@
 # limitations under the License.
 ################################################################################
 
-import sys
-from typing import Any
-
-from data_processing.data_access import DataAccess, DataAccessFactory
-from kfp_support.workflow_support.utils import KFPUtils, RayRemoteJobs
-
-
-def execute_ray_jobs(
-    name: str,  # name of Ray cluster
-    d_access: DataAccess,  # data access
-    additional_params: dict[str, Any],
-    e_params: dict[str, Any],
-    exec_script_name: str,
-    server_url: str,
-) -> None:
-    """
-    Execute Ray job on a cluster periodically printing execution log. Completes when Ray job completes
-    (succeeds or fails)
-    :param name: cluster name
-    :param d_access: data access class
-    :param additional_params: additional parameters for the job
-    :param e_params: job execution parameters (specific for a specific transform,
-                        generated by the transform workflow)
-    :param exec_script_name: script to run (has to be present in the image)
-    :param server_url: API server url
-    :return: None
-    """
-    # get current namespace
-    ns = KFPUtils.get_namespace()
-    if ns == "":
-        print(f"Failed to get namespace")
-        sys.exit(1)
-    # submit job
-    remote_jobs = RayRemoteJobs(
-        server_url=server_url,
-        http_retries=additional_params.get("http_retries", 5),
-        wait_interval=additional_params.get("wait_interval", 2),
-    )
-    status, error, submission = remote_jobs.submit_job(
-        name=name, namespace=ns, request=e_params, executor=exec_script_name
-    )
-    if status != 200:
-        print(f"Failed to submit job - status: {status}, error: {error}, submission id {submission}")
-        exit(1)
-
-    print(f"submitted job successfully, submission id {submission}")
-    # print execution log
-    remote_jobs.follow_execution(
-        name=name,
-        namespace=ns,
-        submission_id=submission,
-        data_access=d_access,
-        print_timeout=additional_params.get("wait_print_tmout", 120),
-        job_ready_timeout=additional_params.get("wait_job_ready_tmout", 600),
-    )
+from data_processing.data_access import DataAccessFactory
+from kfp_support.workflow_support.utils import KFPUtils, execute_ray_job
 
 
 if __name__ == "__main__":
@@ -108,7 +55,7 @@ if __name__ == "__main__":
         "{'access_key': '" + access_key + "', 'secret_key': '" + secret_key + "', 'url': '" + url + "'}"
     )
     # Execute Ray jobs
-    execute_ray_jobs(
+    execute_ray_job(
         name=cluster_name,
         d_access=data_access,
         additional_params=KFPUtils.load_from_json(args.additional_params.replace("'", '"')),
